+++ Sources/TraceLog/TraceLog.swift
@@ -21,23 +21,70 @@ import Swift
 import Foundation
 
 ///
-/// Initializes TraceLog with an optional array of Writers and the Environment.
+/// Initializes TraceLog with an optional mode, array of Writers and the Environment.
 ///
 /// This call is optional but in order to read from the environment on start up,
-/// This method must be called.
+/// this method must be called.
 ///
 /// - Parameters:
+///     - mode:        The `ConcurrencyMode` to run all `Writer`s in. Async is the default.
 ///     - writers:     An Array of objects that implement the Writer protocol used to process messages that are logged. Note the writers are called in the order they are in this array.
 ///     - environment: Either a Dictionary<String, String> or an Environment object that contains the key/value pairs of configuration variables for TraceLog.
 ///
-/// Example:
+/// - Example:
+///
+/// Start TraceLog in the default mode, with default writers.
+/// ```
+///     TraceLog.configure()
+/// ```
+///
+/// Start TraceLog the with default writer in `.direct` mode.
+/// ```
+///     TraceLog.configure(mode: .direct)
+/// ```
+///
+/// Start TraceLog in the default mode, replacing the default writer with `MyWriter` reading the environment for log level settings.
+/// ```
+///     TraceLog.configure(writers: [MyWriter()])
+/// ```
+///
+/// Start TraceLog in the default mode, replacing the default writer with `MyWriter` and setting log levels programmatically.
 /// ```
 ///     TraceLog.configure(writers: [MyWriter()], environment: ["LOG_ALL": "TRACE4",
 ///                                                              "LOG_PREFIX_NS" : "ERROR",
 ///                                                              "LOG_TAG_TraceLog" : "TRACE4"])
 /// ```
 ///
-public func configure(writers: [Writer] = [ConsoleWriter()], environment: Environment = Environment()) {
+public func configure(mode: ConcurrencyMode = .default, writers: [Writer] = [ConsoleWriter()], environment: Environment = Environment()) {
+    #if !TRACELOG_DISABLED
+        Logger.configure(writers: writers.map( { mode.writerMode(for: $0) } ), environment: environment)
+    #endif
+}
+
+///
+/// Initializes TraceLog with an optional array of Writers specifying thier ConcurrencyMode and the Environment.
+///
+/// - Parameters:
+///     - writers:     An Array of `Writers` wrapped in a `WriterConcurrencyMode`.
+///     - environment: Either a Dictionary<String, String> or an Environment object that contains the key/value pairs of configuration variables for TraceLog.
+///
+/// - Example:
+///
+/// Start TraceLog replacing the default writer with `MyWriter` running in `.direct` mode, `MyHTTPWriter`
+/// in `.async` mode, and reading the environment for log level settings.
+/// ```
+///     TraceLog.configure(writers: [.direct(MyWriter()), .async(MyHTTPWriter())])
+/// ```
+///
+/// Start TraceLog replacing the default writer with `MyWriter` running in `.async` mode and  setting log
+/// levels programmatically.
+/// ```
+///     TraceLog.configure(writers: [.async(MyWriter())], environment: ["LOG_ALL": "TRACE4",
+///                                                                     "LOG_PREFIX_NS" : "ERROR",
+///                                                                     "LOG_TAG_TraceLog" : "TRACE4"])
+/// ```
+///
+public func configure(writers: [WriterConcurrencyMode], environment: Environment = Environment()) {
     #if !TRACELOG_DISABLED
         Logger.configure(writers: writers, environment: environment)
     #endif